Restructured component folder to match routes
