+++ app/components/Preferences.tsx
@@ -1,7 +1,6 @@
 'use client'
 import './preferences.scss'
 
-import { useContext } from 'react'
 import { useRouter } from 'next/navigation'
 
 import { useWeaveContext } from '@/app/resources/contexts/weavecontext'
