+++ app/components/preferences.scss
